- changed AnswerType.rawValue type from Int16 to Int - removed useless properties

+++ 8 Ball/BD/AnswerManager.swift
@@ -74,7 +74,7 @@ extension AnswerManager {
         }
         else {
             let answer = Answer(context: context)
-            answer.type = type.rawValue
+            answer.type = Int16(type.rawValue)
             answer.title = answerString
             answer.createdByUser = createdByUser
             return answer
@@ -105,7 +105,7 @@ extension AnswerManager {
             throw AnswerError.answerAlreadyExists
         }
         answer.title = answerTitle
-        answer.type = type.rawValue
+        answer.type = Int16(type.rawValue)
     }
     
     static func deleteAnswer(_ answer: Answer, context: NSManagedObjectContext) throws {
